fastapi-assignment/pyproject.toml
[tool.poetry]
name = "fastapi-assignment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
tortoise-orm = "^0.25.1"
python-multipart = "^0.0.20"
pydantic-settings = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
